% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rayleigh.R
\name{calc_rayleigh}
\alias{calc_rayleigh}
\title{Calculate the Rayleigh Isotope Curve}
\usage{
calc_rayleigh(delta_initial, epsilon)
}
\arguments{
\item{delta_initial}{The initial delta value}

\item{epsilon}{The isotope enrichement factor}
}
\description{
This function calculates the delta values of the residual material
along a Rayleigh curve and returns a data.frame of results.
}
