% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rayleigh.R
\name{plot_rayleigh}
\alias{plot_rayleigh}
\title{Plot the Rayleight Isotope Curve}
\usage{
plot_rayleigh(df)
}
\arguments{
\item{df}{The data.frame returned by calc_rayleigh}
}
\description{
Plot the Rayleigh Isotope Curve
}
\details{
This function takes a simple data.frame, gathers it and plots
the delta values of the residual material along a Rayleigh
curve and returns a ggplot object.
}
